migration file4

+++ database/migrations/2024_03_21_013148_create_grades_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::create('grades', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->integer('id');
             $table->string('name', 255);
             $table->timestamps();
         });
@@ -28,6 +28,6 @@ return new class extends Migration
     public function down()
     {
         Schema::dropIfExists('grades');
-        
+
     }
 };
